+++ docs/downloadCount.data.ts
@@ -1,0 +1,7 @@
+import { fetchNpmJsDownloadCount } from './components/starCount.ts'
+
+export default {
+  async load() {
+    return await fetchNpmJsDownloadCount()
+  },
+}