+++ packages/pglite/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @electric-sql/pglite
 
+## 0.4.5
+
+### Patch Changes
+
+- c6bddde: Fix caching of artifacts such that they are not downloaded multiple times
+
+## 0.4.4
+
+### Patch Changes
+
+- b88c5c3: Disable checkpointer
+
+## 0.4.3
+
+### Patch Changes
+
+- 2ae666f: Default database, user and role are now all "postgres"
+- fb95e66: Allow setting initial memory size.
+- 65fc101: Disable background workers.
+
+## 0.4.2
+
+### Patch Changes
+
+- 41632c4: Allow passing initdb.wasm asset for bundlers that need it.
+
 ## 0.4.1
 
 ### Patch Changes